requirements.txt
langchain[groq]
langgraph
langchain-text-splitters
langchain_community
langchain-huggingface
langchain-qdrant
langchain-google-genai
wikipedia
python_dotenv
beautifulsoup4
requests
faiss-cpu
sentence-transformers
huggingface_hub
# extração PDF toolkit.py
llama-index
pypdf
qdrant-client
streamlit
